IntelliNews App/Backend/requirements.txt
APScheduler==3.10.4
gnews==0.3.7
groq==0.9.0
newspaper3k==0.2.8
pandas==2.0.3
Requests==2.32.3
sentence_transformers==3.0.1
torch==2.5.0.dev20240729
transformers==4.42.4
transformers==4.41.2
